+++ CHANGELOG.md
@@ -1,15 +1,18 @@
-<a name="0.1.2"></a>
-## 0.1.2 (2018-07-30)
+<a name="0.1.3"></a>
+## [0.1.3](https://github.com/roalcantara/factory-bot-ts/compare/v0.1.2...v0.1.3) (2018-08-07)
 
 
 ### Bug Fixes
 
+* **factories:** Fix the clear method ([4d96588](https://github.com/roalcantara/factory-bot-ts/commit/4d96588))
+* **factories:** Fix the count method ([9b29a62](https://github.com/roalcantara/factory-bot-ts/commit/9b29a62))
 * **factories:** Require partial attributes ([22f55c6](https://github.com/roalcantara/factory-bot-ts/commit/22f55c6))
 * **release:** Add dist folder ([8754133](https://github.com/roalcantara/factory-bot-ts/commit/8754133))
 
 
 ### Features
 
+* **factories:** Add extend method ([181d934](https://github.com/roalcantara/factory-bot-ts/commit/181d934))
 * **factories:** Add factory-bot ([631c447](https://github.com/roalcantara/factory-bot-ts/commit/631c447))